requirements.txt
openai>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
pyyaml>=6.0
PyGithub>=1.59.0
scholarly>=1.7.0
selenium>=4.10.0
webdriver-manager>=4.0.0
pylatexenc>=2.10
streamlit>=1.28.0
